Introduced new enumerator for the vertices, i.e. the Python interface will get all three different data types right right from the start.

+++ src/peano4/datamanagement/VertexEnumerator.h
@@ -85,7 +85,7 @@ class peano4::datamanagement::VertexEnumerator {
       tarch::la::Vector<Dimensions,double> result( _x );
       std::bitset<Dimensions> myset(i);
       for (int d=0; d<Dimensions; d++) {
-        result(d) += myset[d] * _h(d);
+        result(d) += static_cast<double>(myset[d]) * _h(d);
       }
       return result;
 	}
